+++ src/main/java/com/imsweb/layout/record/fixed/naaccr/NaaccrLayout.java
@@ -5,12 +5,14 @@ package com.imsweb.layout.record.fixed.naaccr;
 
 import java.io.IOException;
 import java.io.InputStream;
-import java.io.InputStreamReader;
-import java.io.Reader;
 import java.io.StringWriter;
 import java.net.URL;
 import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
+
+import org.apache.commons.io.IOUtils;
 
 import com.imsweb.layout.Field;
 import com.imsweb.layout.LayoutFactory;
@@ -123,6 +125,172 @@ public class NaaccrLayout extends FixedColumnsLayout {
         _DEFAULT_CSS_14_AND_LATER.append(".chap10-para ul li { margin-bottom: 0;}\n");
     }
 
+    // mapping between the NAACCR layout name and the NAACCR XML IDs
+    private static Map<String, String> _XML_TO_LAYOUT_MAPPING = new HashMap<>();
+
+    static {
+        _XML_TO_LAYOUT_MAPPING.put("patientSysIdHospital", "patientSystemIdHosp");
+        _XML_TO_LAYOUT_MAPPING.put("registryIdNpi", "npiRegistryId");
+        _XML_TO_LAYOUT_MAPPING.put("addressAtDxCity", "addrAtDxCity");
+        _XML_TO_LAYOUT_MAPPING.put("addressAtDxState", "addrAtDxState");
+        _XML_TO_LAYOUT_MAPPING.put("addressAtDxCounty", "countyAtDx");
+        _XML_TO_LAYOUT_MAPPING.put("countyDxGeocode1990", "countyAtDxGeocode1990");
+        _XML_TO_LAYOUT_MAPPING.put("countyDxGeocode2000", "countyAtDxGeocode2000");
+        _XML_TO_LAYOUT_MAPPING.put("countyDxGeocode2010", "countyAtDxGeocode2010");
+        _XML_TO_LAYOUT_MAPPING.put("countyDxGeocode2020", "countyAtDxGeocode2020");
+        _XML_TO_LAYOUT_MAPPING.put("addressAtDxPostalCode", "addrAtDxPostalCode");
+        _XML_TO_LAYOUT_MAPPING.put("addressAtDxCountry", "addrAtDxCountry");
+        _XML_TO_LAYOUT_MAPPING.put("censusTract708090", "censusTract19708090");
+        _XML_TO_LAYOUT_MAPPING.put("censusCodingSys708090", "censusCodSys19708090");
+        _XML_TO_LAYOUT_MAPPING.put("censusPovertyIndictr", "censusTrPovertyIndictr");
+        _XML_TO_LAYOUT_MAPPING.put("nhia", "nhiaDerivedHispOrigin");
+        _XML_TO_LAYOUT_MAPPING.put("ihs", "ihsLink");
+        _XML_TO_LAYOUT_MAPPING.put("napiia", "raceNapiia");
+        _XML_TO_LAYOUT_MAPPING.put("ageAtDx", "ageAtDiagnosis");
+        _XML_TO_LAYOUT_MAPPING.put("birthDate", "dateOfBirth");
+        _XML_TO_LAYOUT_MAPPING.put("birthDateFlag", "dateOfBirthFlag");
+        _XML_TO_LAYOUT_MAPPING.put("birthPlace", "birthplace");
+        _XML_TO_LAYOUT_MAPPING.put("occupationCodeCensus", "censusOccCode19702000");
+        _XML_TO_LAYOUT_MAPPING.put("industryCodeCensus", "censusIndCode19702000");
+        _XML_TO_LAYOUT_MAPPING.put("occupationIndustryCodingSys", "censusOccIndSys7000");
+        _XML_TO_LAYOUT_MAPPING.put("censusCertainty708090", "censusTrCert19708090");
+        _XML_TO_LAYOUT_MAPPING.put("censusCertainty2000", "censusTrCertainty2000");
+        _XML_TO_LAYOUT_MAPPING.put("gisQuality", "gisCoordinateQuality");
+        _XML_TO_LAYOUT_MAPPING.put("censusCertainty2010", "censusTrCertainty2010");
+        _XML_TO_LAYOUT_MAPPING.put("censusBlockGroup708090", "censusBlockGrp197090");
+        _XML_TO_LAYOUT_MAPPING.put("morphologyIcdO2", "morphTypebehavIcdO2");
+        _XML_TO_LAYOUT_MAPPING.put("dateOfMultipleTumorsFlag", "dateOfMultTumorsFlag");
+        _XML_TO_LAYOUT_MAPPING.put("multiTumorRptAsOnePrim", "multTumRptAsOnePrim");
+        _XML_TO_LAYOUT_MAPPING.put("dateOfMultipleTumors", "dateOfMultTumors");
+        _XML_TO_LAYOUT_MAPPING.put("gradePathSys", "gradePathSystem");
+        _XML_TO_LAYOUT_MAPPING.put("morphCodingSysOriginal", "morphCodingSysOriginl");
+        _XML_TO_LAYOUT_MAPPING.put("morphologyIcdO3", "morphTypebehavIcdO3");
+        _XML_TO_LAYOUT_MAPPING.put("histologyIcdO3", "histologicTypeIcdO3");
+        _XML_TO_LAYOUT_MAPPING.put("behaviorIcdO3", "behaviorCodeIcdO3");
+        _XML_TO_LAYOUT_MAPPING.put("reportingFacilityNpi", "npiReportingFacility");
+        _XML_TO_LAYOUT_MAPPING.put("sequenceNumberHosp", "sequenceNumberHospital");
+        _XML_TO_LAYOUT_MAPPING.put("extentOfDisease10Digit", "extentOfDisease10Dig");
+        _XML_TO_LAYOUT_MAPPING.put("eodCodingSys", "codingSystemForEod");
+        _XML_TO_LAYOUT_MAPPING.put("metsDxBone", "metsAtDxBone");
+        _XML_TO_LAYOUT_MAPPING.put("metsDxBrain", "metsAtDxBrain");
+        _XML_TO_LAYOUT_MAPPING.put("metsDxDistantLn", "metsAtDxDistantLn");
+        _XML_TO_LAYOUT_MAPPING.put("metsDxLiver", "metsAtDxLiver");
+        _XML_TO_LAYOUT_MAPPING.put("metsDxLung", "metsAtDxLung");
+        _XML_TO_LAYOUT_MAPPING.put("metsDxOther", "metsAtDxOther");
+        _XML_TO_LAYOUT_MAPPING.put("pediatricStagingSys", "pediatricStagingSystem");
+        _XML_TO_LAYOUT_MAPPING.put("dateOfInitialRx", "dateInitialRxSeer");
+        _XML_TO_LAYOUT_MAPPING.put("dateOfInitialRxFlag", "dateInitialRxSeerFlag");
+        _XML_TO_LAYOUT_MAPPING.put("dateOf1stCrsRxCoc", "date1stCrsRxCoc");
+        _XML_TO_LAYOUT_MAPPING.put("dateOf1stCrsRxCocFlag", "date1stCrsRxCocFlag");
+        _XML_TO_LAYOUT_MAPPING.put("rxSummSurgicalApproach", "rxSummSurgicalApproch");
+        _XML_TO_LAYOUT_MAPPING.put("rxCodingSysCurrent", "rxCodingSystemCurrent");
+        _XML_TO_LAYOUT_MAPPING.put("rxSummSystemicSurgSeq", "rxSummSystemicSurSeq");
+        _XML_TO_LAYOUT_MAPPING.put("subsqRx2ndCourseDateFlag", "subsqRx2ndcrsDateFlag");
+        _XML_TO_LAYOUT_MAPPING.put("subsqRx3rdCourseDateFlag", "subsqRx3rdcrsDateFlag");
+        _XML_TO_LAYOUT_MAPPING.put("subsqRx4thCourseDateFlag", "subsqRx4thcrsDateFlag");
+        _XML_TO_LAYOUT_MAPPING.put("survDateActiveFollowUp", "survDateActiveFollowup");
+        _XML_TO_LAYOUT_MAPPING.put("survFlagActiveFollowUp", "survFlagActiveFollowup");
+        _XML_TO_LAYOUT_MAPPING.put("survMonthsActiveFollowUp", "survMosActiveFollowup");
+        _XML_TO_LAYOUT_MAPPING.put("survMonthsPresumedAlive", "survMosPresumedAlive");
+        _XML_TO_LAYOUT_MAPPING.put("addressCurrentCity", "addrCurrentCity");
+        _XML_TO_LAYOUT_MAPPING.put("addressCurrentState", "addrCurrentState");
+        _XML_TO_LAYOUT_MAPPING.put("addressCurrentPostalCode", "addrCurrentPostalCode");
+        _XML_TO_LAYOUT_MAPPING.put("addressCurrentCountry", "addrCurrentCountry");
+        _XML_TO_LAYOUT_MAPPING.put("addressCurrentCounty", "countyCurrent");
+        _XML_TO_LAYOUT_MAPPING.put("primarySiteIcdO1", "siteIcdO1");
+        _XML_TO_LAYOUT_MAPPING.put("morphologyIcdO1", "morphIcdO1");
+        _XML_TO_LAYOUT_MAPPING.put("overRideSsNodesPos", "overRideSsNodespos");
+        _XML_TO_LAYOUT_MAPPING.put("overRideHospSeqDxConf", "overRideHospseqDxconf");
+        _XML_TO_LAYOUT_MAPPING.put("overRideHospSeqSite", "overRideHospseqSite");
+        _XML_TO_LAYOUT_MAPPING.put("overRideSiteTnmStgGrp", "overRideSiteTnmStggrp");
+        _XML_TO_LAYOUT_MAPPING.put("overRideSeqNoDxConf", "overRideSeqnoDxconf");
+        _XML_TO_LAYOUT_MAPPING.put("overRideSiteLatSeqNo", "overRideSiteLatSeqno");
+        _XML_TO_LAYOUT_MAPPING.put("overRideSurgDxConf", "overRideSurgDxconf");
+        _XML_TO_LAYOUT_MAPPING.put("overRideSiteEodDxDate", "overRideSiteEodDxDt");
+        _XML_TO_LAYOUT_MAPPING.put("addressAtDxStreetName", "addrAtDxNoStreet");
+        _XML_TO_LAYOUT_MAPPING.put("addressAtDxSupplementl", "addrAtDxSupplementl");
+        _XML_TO_LAYOUT_MAPPING.put("addressCurrentStreetName", "addrCurrentNoStreet");
+        _XML_TO_LAYOUT_MAPPING.put("addressCurrentSupplementl", "addrCurrentSupplementl");
+        _XML_TO_LAYOUT_MAPPING.put("followUpContactNoAndStreet", "followUpContactNost");
+        _XML_TO_LAYOUT_MAPPING.put("institutionReferredFromNpi", "npiInstReferredFrom");
+        _XML_TO_LAYOUT_MAPPING.put("institutionReferredToNpi", "npiInstReferredTo");
+        _XML_TO_LAYOUT_MAPPING.put("followingRegistryNpi", "npiFollowingRegistry");
+        _XML_TO_LAYOUT_MAPPING.put("physicianManagingNpi", "npiPhysicianManaging");
+        _XML_TO_LAYOUT_MAPPING.put("physicianFollowUpNpi", "npiPhysicianFollowUp");
+        _XML_TO_LAYOUT_MAPPING.put("physicianPrimarySurgNpi", "npiPhysicianPrimarySurg");
+        _XML_TO_LAYOUT_MAPPING.put("physician3Npi", "npiPhysician3");
+        _XML_TO_LAYOUT_MAPPING.put("physician4Npi", "npiPhysician4");
+        _XML_TO_LAYOUT_MAPPING.put("rxTextRadiationBeam", "rxTextRadiation");
+        _XML_TO_LAYOUT_MAPPING.put("placeOfDiagnosis", "textPlaceOfDiagnosis");
+        _XML_TO_LAYOUT_MAPPING.put("csVersionOriginal", "csVersionInputOriginal");
+        _XML_TO_LAYOUT_MAPPING.put("derivedAjcc6TDescriptor", "derivedAjcc6TDescript");
+        _XML_TO_LAYOUT_MAPPING.put("derivedAjcc6NDescriptor", "derivedAjcc6NDescript");
+        _XML_TO_LAYOUT_MAPPING.put("derivedAjcc6MDescriptor", "derivedAjcc6MDescript");
+        _XML_TO_LAYOUT_MAPPING.put("derivedAjcc6StageGroup", "derivedAjcc6StageGrp");
+        _XML_TO_LAYOUT_MAPPING.put("archiveFinNpi", "npiArchiveFin");
+        _XML_TO_LAYOUT_MAPPING.put("ruralUrbanContinuum1993", "ruralurbanContinuum1993");
+        _XML_TO_LAYOUT_MAPPING.put("ruralUrbanContinuum2003", "ruralurbanContinuum2003");
+        _XML_TO_LAYOUT_MAPPING.put("ruralUrbanContinuum2013", "ruralurbanContinuum2013");
+        _XML_TO_LAYOUT_MAPPING.put("derivedAjcc7TDescriptor", "derivedAjcc7TDescript");
+        _XML_TO_LAYOUT_MAPPING.put("derivedAjcc7NDescriptor", "derivedAjcc7NDescript");
+        _XML_TO_LAYOUT_MAPPING.put("derivedAjcc7MDescriptor", "derivedAjcc7MDescript");
+        _XML_TO_LAYOUT_MAPPING.put("derivedAjcc7StageGroup", "derivedAjcc7StageGrp");
+        _XML_TO_LAYOUT_MAPPING.put("derivedPreRx7T", "derivedPrerx7T");
+        _XML_TO_LAYOUT_MAPPING.put("derivedPreRx7TDescriptor", "derivedPrerx7TDescrip");
+        _XML_TO_LAYOUT_MAPPING.put("derivedPreRx7N", "derivedPrerx7N");
+        _XML_TO_LAYOUT_MAPPING.put("derivedPreRx7NDescriptor", "derivedPrerx7NDescrip");
+        _XML_TO_LAYOUT_MAPPING.put("derivedPreRx7M", "derivedPrerx7M");
+        _XML_TO_LAYOUT_MAPPING.put("derivedPreRx7MDescriptor", "derivedPrerx7MDescrip");
+        _XML_TO_LAYOUT_MAPPING.put("derivedPreRx7StageGroup", "derivedPrerx7StageGrp");
+        _XML_TO_LAYOUT_MAPPING.put("derivedPostRx7T", "derivedPostrx7T");
+        _XML_TO_LAYOUT_MAPPING.put("derivedPostRx7N", "derivedPostrx7N");
+        _XML_TO_LAYOUT_MAPPING.put("derivedPostRx7M", "derivedPostrx7M");
+        _XML_TO_LAYOUT_MAPPING.put("derivedPostRx7StageGroup", "derivedPostrx7StgeGrp");
+        _XML_TO_LAYOUT_MAPPING.put("derivedNeoAdJuvRxFlag", "derivedNeoadjuvRxFlag");
+        _XML_TO_LAYOUT_MAPPING.put("derivedSeerCombinedStgGrp", "derivedSeerCmbStgGrp");
+        _XML_TO_LAYOUT_MAPPING.put("derivedSeerCombinedTSrc", "derivedSeerCmbTSrc");
+        _XML_TO_LAYOUT_MAPPING.put("derivedSeerCombinedNSrc", "derivedSeerCmbNSrc");
+        _XML_TO_LAYOUT_MAPPING.put("derivedSeerCombinedMSrc", "derivedSeerCmbMSrc");
+        _XML_TO_LAYOUT_MAPPING.put("seerSiteSpecificFactor1", "seerSiteSpecificFact1");
+        _XML_TO_LAYOUT_MAPPING.put("seerSiteSpecificFactor2", "seerSiteSpecificFact2");
+        _XML_TO_LAYOUT_MAPPING.put("seerSiteSpecificFactor3", "seerSiteSpecificFact3");
+        _XML_TO_LAYOUT_MAPPING.put("seerSiteSpecificFactor4", "seerSiteSpecificFact4");
+        _XML_TO_LAYOUT_MAPPING.put("seerSiteSpecificFactor5", "seerSiteSpecificFact5");
+        _XML_TO_LAYOUT_MAPPING.put("seerSiteSpecificFactor6", "seerSiteSpecificFact6");
+        _XML_TO_LAYOUT_MAPPING.put("overrideCs1", "overRideCs1");
+        _XML_TO_LAYOUT_MAPPING.put("overrideCs2", "overRideCs2");
+        _XML_TO_LAYOUT_MAPPING.put("overrideCs3", "overRideCs3");
+        _XML_TO_LAYOUT_MAPPING.put("overrideCs4", "overRideCs4");
+        _XML_TO_LAYOUT_MAPPING.put("overrideCs5", "overRideCs5");
+        _XML_TO_LAYOUT_MAPPING.put("overrideCs6", "overRideCs6");
+        _XML_TO_LAYOUT_MAPPING.put("overrideCs7", "overRideCs7");
+        _XML_TO_LAYOUT_MAPPING.put("overrideCs8", "overRideCs8");
+        _XML_TO_LAYOUT_MAPPING.put("overrideCs9", "overRideCs9");
+        _XML_TO_LAYOUT_MAPPING.put("overrideCs10", "overRideCs10");
+        _XML_TO_LAYOUT_MAPPING.put("overrideCs11", "overRideCs11");
+        _XML_TO_LAYOUT_MAPPING.put("overrideCs12", "overRideCs12");
+        _XML_TO_LAYOUT_MAPPING.put("overrideCs13", "overRideCs13");
+        _XML_TO_LAYOUT_MAPPING.put("overrideCs14", "overRideCs14");
+        _XML_TO_LAYOUT_MAPPING.put("overrideCs15", "overRideCs15");
+        _XML_TO_LAYOUT_MAPPING.put("overrideCs16", "overRideCs16");
+        _XML_TO_LAYOUT_MAPPING.put("overrideCs17", "overRideCs17");
+        _XML_TO_LAYOUT_MAPPING.put("overrideCs18", "overRideCs18");
+        _XML_TO_LAYOUT_MAPPING.put("overrideCs19", "overRideCs19");
+        _XML_TO_LAYOUT_MAPPING.put("overrideCs20", "overRideCs20");
+        _XML_TO_LAYOUT_MAPPING.put("pathReportFacId1", "pathReportingFacId1");
+        _XML_TO_LAYOUT_MAPPING.put("pathReportFacId2", "pathReportingFacId2");
+        _XML_TO_LAYOUT_MAPPING.put("pathReportFacId3", "pathReportingFacId3");
+        _XML_TO_LAYOUT_MAPPING.put("pathReportFacId4", "pathReportingFacId4");
+        _XML_TO_LAYOUT_MAPPING.put("pathReportFacId5", "pathReportingFacId5");
+    }
+
+    /**
+     * Utility method to translate a layout field name into the corresponding NAACCR XML ID. Supports NAACCR 18 and later. Returns the input name if not found.
+     */
+    public static String getXmlIdFromLayoutName(String name) {
+        return _XML_TO_LAYOUT_MAPPING.getOrDefault(name, name);
+    }
+
     protected String _naaccrVersion;
 
     protected String _majorNaaccrVersion;
@@ -230,24 +398,9 @@ public class NaaccrLayout extends FixedColumnsLayout {
         if (docPath == null)
             return null;
 
-        try {
-            InputStream is = docPath.openStream();
-            if (is == null)
-                return null;
-
-            StringWriter writer = new StringWriter();
-            Reader reader = new InputStreamReader(is, "UTF-8");
-
-            char[] buf = new char[1024];
-            int n;
-            while ((n = reader.read(buf, 0, 1024)) != -1)
-                writer.write(buf, 0, n);
-            writer.flush();
-            writer.close();
-            reader.close();
-
+        try (InputStream is = docPath.openStream(); StringWriter writer = new StringWriter()) {
+            IOUtils.copy(is, writer, "UTF-8");
             result = writer.toString();
-
         }
         catch (IOException e) {
             /* do nothing, result will be null, as per specs */